requirements.txt
python-dotenv
langchain
openai
browser-use
asyncio
playwright